refactor: phase 6 hexagonal polish
- Fixed the dependency injection leak in `WorkspaceManager` by removing the `LocalFileLoader` default and injecting it explicitly from `cli/main.py` instead.
- Enforced port contracts on the `Langchain` and `MCP` LLM providers.
- Decentralized exceptions by extracting `LineageIntegrityError` to a specific `core/workflow/exceptions.py` file, exporting it through the `core/workflow` facade, and updating corresponding imports across the repository.
