+++ src/coreason_manifest/core/telemetry/request.py
@@ -3,8 +3,9 @@ from uuid import uuid4
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from coreason_manifest.core.common.exceptions import LineageIntegrityError, ManifestError, ManifestErrorCode
+from coreason_manifest.core.common.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.core.common.identity import SessionContext
+from coreason_manifest.core.workflow import LineageIntegrityError
 from coreason_manifest.core.workflow.flow import GraphFlow, LinearFlow
