+++ src/coreason_manifest/core/workflow/exceptions.py
@@ -1,0 +1,23 @@
+from coreason_manifest.core.common.exceptions import (
+    FaultSeverity,
+    ManifestError,
+    ManifestErrorCode,
+    RecoveryAction,
+    SemanticFault,
+)
+
+
+class LineageIntegrityError(ManifestError):
+    """
+    Raised when a trace lineage violation is detected.
+    """
+
+    def __init__(self, message: str) -> None:
+        super().__init__(
+            SemanticFault(
+                error_code=ManifestErrorCode.SEC_LINEAGE_001,
+                message=message,
+                severity=FaultSeverity.CRITICAL,
+                recovery_action=RecoveryAction.HALT,
+            )
+        )
